nitpick: portability have{Binary,Callable}
slightly more portable, per the following paragraph:

> Now, on the standard front, POSIX specifies the command -v and -V
> commands (which used to be optional until POSIX.2008). UNIX specifies
> the type command (no option). That's all (where, which, whence are not
> specified in any standard).

found at https://unix.stackexchange.com/questions/85249/why-not-use-which-what-to-use-then/85250#85250

"portability" here is obviously kind of meaningless in the case of yabashlib,
which is a Bash project. But it's useful for me in case I ever want to come
here and use this as a reference source and copy/paste the implementation to
somewhere else.